Scratch Codes for Supervised and Unsupervised Learning
This Python notebook serves as a comprehensive guide to implementing various machine learning algorithms from scratch using Python. In this notebook, we cover a wide range of machine learning concepts, including supervised and unsupervised learning algorithms, as well as techniques for data preprocessing, model evaluation, and visualization. Each algorithm is implemented step-by-step, providing insights into the underlying principles and algorithms' inner workings. Moreover, visualizations are incorporated to facilitate better understanding and intuition behind the algorithms. Whether you're a beginner looking to grasp the fundamentals of machine learning or an enthusiast aiming to deepen your understanding, this notebook offers a practical and hands-on approach to learning machine learning algorithms in Python.